Allow coverage.py to be executed by pointing Python at a repository's directory.
Since Python 2.6 one can specify a __main__.py in a directory to tell
Python what to do if it is pointed at simply a directory. For
instance, assuming one has a clone of coverage.py sitting at
../coveragepy, this patch allows for::

  python ../coveragepy run ...

This is extremely handy if you want to use coverage.py straight from a
clone without adding the clone's directory to sys.path (which can be
an issue at least in the stdlib as that will pick up 'test' as a
package, masking the stdlib's own 'test' package).