Combine no longer appends by default
Combine used to always load an existing .coverage file. This lead to confusing
results and extra tox-clean steps.  Now the default is to not load the existing
file, though a new --append switch on coverage combine gets you that behavior
if you need it.

This also pointed up an issue with concurrency=multiprocessing, which is that
the child processes automatically used parallel=True, but the parent process
did not.  Now concurrency=multiprocessing implies parallel=True.