html total must account for files 100% covered #549
The --skip-covered implementation for html must not just skip the
files that are 100% covered. It must remember them and not fail with a
"no coverage data" when all files are 100% covered.

The statistics about the files that are 100% covered must also be
summed in the total for the index.html page.

close #549

--HG--
branch : issue-549